requirements.txt
fastapi[all]
sqlalchemy==2.0.20
alembic==1.12.0
asyncpg==0.28.0
passlib==1.7.4
python-jose==3.3.0
fastapi-cache2[redis]
celery==5.3.4
cached flower==2.0.1
Pillow==10.0.0
sqladmin==0.14.1
pytest==7.4.2
black==23.9.1
flake8==6.1.0
autoflake==2.2.1
isort==5.12.0
pyright==1.1.327